agent/requirements.txt
livekit-agents[google,images]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
asyncio
